.archive/.old/backend/requirements.txt
# FastAPI 核心
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# WebSocket
websockets==13.1

# Celery 任务队列
celery[redis]==5.4.0
redis==5.2.0

# AI/ML 依赖
ultralytics>=8.3.0
torch>=2.4.0
torchvision>=0.19.0
tensorflow>=2.17.0
onnx>=1.16.0

# 文件处理
python-multipart==0.0.12
aiofiles==24.1.0

# 工具库
structlog==24.4.0
loguru>=0.7.0
python-dotenv==1.0.1
httpx==0.28.0
pyyaml==6.0.2

# 开发工具（可选）
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0
black==24.10.0
ruff==0.8.0
mypy==1.13.0
